configurator timing spec

+++ c/f2_atomic.h
@@ -192,7 +192,7 @@ static inline int atomic_inc_and_test( atomic_t *v ) {
  */
 static inline int atomic_add_negative( int i, atomic_t *v ) {
 #if GCC_ATOMIC_BUILTINS
-  return (__sync_add_and_fetch(&v->counter, i) < 0);
+  return (((int)(__sync_add_and_fetch(&v->counter, i))) < 0);
 #else
   v->counter ++;
   return (v->counter < 0);
